[py visualization] Split out flaky visualizer reload test
